chore(ci): make the Go lint gate real and correct stale ledger evidence
Two findings from reconciling the port contract matrix against the integrated HEAD.

The `Lint` step of the Go CI job never ran the configured linters: `make lint`
silently falls back to `go vet` when `golangci-lint` is absent, and no workflow
step installs it, so the `.golangci.yml` configuration (unused, ineffassign,
staticcheck, unconvert) was never enforced in CI. Running it locally surfaced
four issues, all in Go oracles and test helpers, none in product code:

* an unnecessary `uint32` conversion in the poll-inbox oracle,
* two unused struct fields whose containing test double only needs its method to
  fail,
* a struct literal that staticcheck wants as a conversion between two identical
  shapes.

Fixing them and making `make lint` fail on a CI runner when the linter is missing
keeps the configured gate honest instead of quietly degrading to `go vet`. The
local fallback for development machines without `golangci-lint` is unchanged.

Separately, five CLI rows pointed at `.snap` and `.json` file names that exist
nowhere in the repository. The behaviour they claim is in fact pinned, by named
cases in `rust-tests/parity/cases/cli/behavior.json`; the rows now cite the real
fixture and the categories that put those cases in the byte-exact group.

Verified after the refactor: the poll-inbox and email oracle fixtures are
byte-identical to their committed versions, so no recorded behaviour moved.

+++ rust-tests/parity/oracle/mcp-poll-inbox/main.go
@@ -132,7 +132,7 @@ func (d *scriptedDialer) Dial(_ context.Context, _ email.IMAPConfig) (email.IMAP
 func uidBounds(uidRange string) (uint32, uint32) {
 	parts := strings.SplitN(uidRange, ":", 2)
 	low := uint32(1)
-	high := uint32(^uint32(0))
+	high := ^uint32(0)
 	if len(parts) > 0 && parts[0] != "*" && parts[0] != "" {
 		if value, err := strconv.ParseUint(parts[0], 10, 32); err == nil {
 			low = uint32(value)
